requirements.txt
# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Transformers (Compatible versions)
transformers==4.30.2
tokenizers==0.13.3
huggingface-hub==0.16.4

# Traditional ML
scikit-learn==1.2.2
xgboost==1.7.6
lightgbm==3.3.5
catboost==1.2

# Data Processing
numpy==1.23.5
pandas==1.5.3
opencv-python==4.7.0.72
Pillow==9.5.0
albumentations==1.3.0

# Visualization & Explainability
matplotlib==3.7.1
seaborn==0.12.2
shap==0.41.0
lime==0.2.0.1

# Utilities
tqdm==4.65.0
optuna==3.2.0
joblib==1.2.0
imbalanced-learn==0.10.1

# Hugging Face Datasets
datasets==2.12.0

# Additional
packaging==23.1
pyyaml==6.0